+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - Docker - Fix lumber version in `Dockerfile`.
+- Technical - database analyzer refacto - separate each database analyzer in its own file and add errors.
 
 ### Fixed
 - Command Generate - Fixed error when analyzing MongoDB views.